Hide mirrord_progress::Progress implementations behind a crate feature
