+++ changelog.d/+hide-progress-implementations.internal.md
@@ -1,0 +1,1 @@
+`mirrord_progress::Progress` implementations are now hidden behind the `implementations` feature.
